refactor: own bias-weight for each neuron + bias learns
Each neuron now has it's own bias weight instead of one for each layer.
The bias itself is still 1.0, but the weight that 1.0 is multiplied with
can be different for each neuron. This change was made because otherwise
it wasn't possible to model a xor-gate.
Also the bias-weight will now be adjusted as well, just like the normal
weights (In other words it "learns")