refactor: more efficient training
Did a calculation 6 times instead of once and saving it. Now the result is
calculated once an then saved for further usage.
Did the calculation with Array.map() first and thought there was a bug when
doing it 6 times on the same array, but .map() always creates a new array,
leaving the old one unmodified