feat: extremely large xor network for testing
Added a network that is extremely large (for an xor) to test if the
training algorithm works for several layers with more than two neurons.
The test was successful, but it took a few thousand (think more than 10k)
to learn, since the networks size is not suited for an xor.
(Imporant point to take away for later, more sophisticated problems)